requirements.txt
# API and Data sources
alpaca-py>=0.8.0
alpaca-trade-api>=3.0.0
polygon-api-client>=1.0.0
yfinance>=0.2.12

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=10.0.0
httpx>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ML libraries
scikit-learn==1.2.0
torch==2.0.0
transformers==4.38.0
sentence-transformers==2.2.0
xgboost==2.0.0  # Gradient boosting library

# Database
redis>=4.5.0
aioredis>=2.0.0
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.12.0  # For database migrations

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
asyncio>=3.4.3
aiohttp>=3.8.0
APScheduler>=3.10.0

# Dashboard
dash>=2.9.0
dash-bootstrap-components>=1.4.0
plotly>=5.14.0

# Monitoring and logging
prometheus-client>=0.16.0

# Testing
pytest>=8.0.0
pylint>=3.0.0
